Add a setting to cancel live jobs upon change
Current implementation of live mode will try to schedule jobs
"sequentially":

- the first change causes a job to be scheduled
- we stop listening for changes until the job is finished
- if the job wasn't for the latest change, a new job is scheduled, etc.

This often results in having to wait for the intermediate results to
complete, and is not always convenient.

This change adds a setting, which, if toggled on, instructs Live mode to
clear the queue and interrupt any running jobs upon receiving a new change,
thus avoiding having to wait for intermediate results.