Add SpinBox to Slider settings for easier input
It isn't always convenient to use only sliders for input, especially for
settings with high granularity such as CFG. This change modifies
SliderSetting widget to have a spinbox instead of a simple label next to
the slider. Also, The formatting logic for %'s and x's at the end is
replaced by the suffix logic coming built-in with QSpinBox.